backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.2
Markdown==3.3.3
pylint==2.6.0 
Pillow==8.1.0
django-filter==2.4.0
django-cors-headers==3.6.0